Refactor tests
Avoid too many repeatations
