+++ package.json
@@ -21,11 +21,11 @@
     "swagger-client": "2.1.9-M1"
   },
   "devDependencies": {
-    "chai": "^1.10.0",
-    "cors": "^2.1.1",
-    "docco": "^0.4.0",
+    "chai": "^2.1.0",
+    "cors": "^2.5.3",
+    "docco": "^0.7.0",
     "event-stream": "^3.2.2",
-    "express": "^3.20.0",
+    "express": "^4.12.0",
     "gulp": "^3.8.11",
     "gulp-clean": "^0.3.1",
     "gulp-coffee": "^2.3.1",
@@ -33,16 +33,16 @@
     "gulp-connect": "^2.2.0",
     "gulp-declare": "^0.3.0",
     "gulp-handlebars": "^3.0.1",
-    "gulp-header": "1.2.2",
+    "gulp-header": "^1.2.2",
     "gulp-less": "^3.0.1",
     "gulp-order": "^1.1.1",
     "gulp-rename": "^1.2.0",
     "gulp-uglify": "^1.1.0",
     "gulp-util": "^3.0.4",
     "gulp-watch": "^4.1.1",
-    "gulp-wrap": "^0.10.1",
+    "gulp-wrap": "^0.11.0",
     "http-server": "^0.7.5",
-    "less": "^1.4.2",
+    "less": "^2.4.0",
     "mocha": "^2.1.0",
     "selenium-webdriver": "^2.45.0"
   }
