+++ test/e2e/driver.js
@@ -1,0 +1,9 @@
+/*
+ * Web driver manager
+*/
+
+var webdriver = require('selenium-webdriver');
+
+var driver = new webdriver.Builder().withCapabilities(webdriver.Capabilities.firefox()).build();
+
+module.exports = driver;
